+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 * Go back to user's accounts when clicking on cancel or `x` from fee/fine form. Fixes UIU-1412.
 * Reset patronBlocks before refetching. Fixes UIU-1430 and UIU-1431.
 * Fix bug with wrong displaying of address type. Refs UIU-1404
-* Display `Overdue fine policy` and `Lost item fee policy` to `Loan Details` page. Refs UIU-1246.
+* Display `Overdue fine policy` and `Lost item fee policy` on the `Loan Details` page. Refs UIU-1246.
 
 ## [2.26.0](https://github.com/folio-org/ui-users/tree/v2.26.0) (2019-12-05)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.25.3...v2.26.0)